fix: stop multiple application of wounded conditions (#1020)
This update fixes a problem when applying the wounded condition.  Duplicate conditions would be created for each player.  Now, the condition is applied only once.   Also, please note that the CUB module is incompatible with this Twodsix feature.  So please only enable one of them.  Thanks to @TheLaundry and @comedian999 for finding the issue.