feat: add dice-tray support (#1424)
This small release adds roll support for dice-tray.  To use, install the module, select a token, and press the d20 icon in the roll sidebar.  A pop-up should launch with buttons to add formulas for characteristics and skills for the selected actor. Note that the module hasn't been updated to v11, hence it will show depreciation warnings in the console.  But the module still functions as intended.  This method of rolling isn't as robust as rolling from the actor sheet or using an item macro.  For instance, there is no ability to do chain or contested rolls through the roll pop-up.

The release also includes some ES localization update from @ForjaSalvaje.